refactor(app): 重命名应用类为QVTApplication并更新引用
应用类已重命名为QVTApplication，以更好地反映其作为QQ版本跟踪应用的身份。所有相关引用均已更新，以确保一致性。此外，调整了AndroidManifest.xml中的应用名称引用，并在全项目中替换了Application实例的引用为QVTApplication实例。
