+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
     </queries>
 
     <application
-        android:name=".Application"
+        android:name=".QVTApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:enableOnBackInvokedCallback="true"
