+++ app/src/main/java/com/xiaoniu/qqversionlist/QVTApplication.kt
@@ -21,9 +21,9 @@ package com.xiaoniu.qqversionlist
 import android.app.Application
 import com.google.android.material.color.DynamicColors
 
-class Application : Application() {
+class QVTApplication : Application() {
     companion object {
-        lateinit var instance: com.xiaoniu.qqversionlist.Application
+        lateinit var instance: com.xiaoniu.qqversionlist.QVTApplication
 
         const val SHIPLY_DEFAULT_APPID = "537230561"
         const val SHIPLY_DEFAULT_SDK_VERSION = "1.3.36-RC01"
