+++ app/src/main/java/com/xiaoniu/qqversionlist/ui/ShiplyAdvancedConfigSheetFragment.kt
@@ -31,8 +31,8 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.xiaoniu.qqversionlist.R
-import com.xiaoniu.qqversionlist.Application.Companion.SHIPLY_DEFAULT_APPID
-import com.xiaoniu.qqversionlist.Application.Companion.SHIPLY_DEFAULT_SDK_VERSION
+import com.xiaoniu.qqversionlist.QVTApplication.Companion.SHIPLY_DEFAULT_APPID
+import com.xiaoniu.qqversionlist.QVTApplication.Companion.SHIPLY_DEFAULT_SDK_VERSION
 import com.xiaoniu.qqversionlist.databinding.BottomsheetShiplyAdvancedConfigBinding
 import com.xiaoniu.qqversionlist.util.DataStoreUtil
 import java.util.Locale
