+++ app/src/main/res/layout/item_tim_version_detail.xml
@@ -136,7 +136,6 @@
             android:id="@+id/tv_tim_desc"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:paddingBottom="6dp"
             android:textColor="?attr/colorSecondary"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/tv_tim_detail_date"
